Revert "Merge remote-tracking branch 'origin/master' into 6.3.x"
This reverts commit b689f1c0598e0b279da00bc11bb32a85912d999a, reversing
changes made to 221bf57cc1a45520b3f5466faf597583371b909c.

+++ www/modules/email/controller/AccountController.php
@@ -84,14 +84,14 @@ class AccountController extends \GO\Base\Controller\AbstractModelController {
 		$response['data']['name'] = $alias->name;
 		$response['data']['signature'] = $alias->signature;
 
-//		if (\GO::modules()->isInstalled('addressbook')) {
-			$defaultTemplateModel = \GO\Email\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
+		if (\GO::modules()->isInstalled('addressbook')) {
+			$defaultTemplateModel = \GO\Addressbook\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
 			if ($defaultTemplateModel) {
 				$response['data']['default_account_template_id'] = $defaultTemplateModel->template_id;
 			} else {
 				$response['data']['default_account_template_id'] = '';
 			}
-//		}
+		}
 
 		return parent::afterLoad($response, $model, $params);
 	}
@@ -129,15 +129,15 @@ class AccountController extends \GO\Base\Controller\AbstractModelController {
 			$alias->save();
 		}
 
-		if ( isset($params['default_account_template_id'])) {
+		if (\GO::modules()->addressbook && isset($params['default_account_template_id'])) {
 			if ($params['default_account_template_id']==-1 || empty($params['default_account_template_id'])) {
-				$defaultTemplateModel = \GO\Email\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
+				$defaultTemplateModel = \GO\Addressbook\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
 				if ($defaultTemplateModel)
 					$defaultTemplateModel->delete();
 			} elseif ($params['default_account_template_id']>0) {
-				$defaultTemplateModel = \GO\Email\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
+				$defaultTemplateModel = \GO\Addressbook\Model\DefaultTemplateForAccount::model()->findByPk($model->id);
 				if (!$defaultTemplateModel) {
-					$defaultTemplateModel = new \GO\Email\Model\DefaultTemplateForAccount();
+					$defaultTemplateModel = new \GO\Addressbook\Model\DefaultTemplateForAccount();
 					$defaultTemplateModel->account_id = $model->id;
 				}
 				$defaultTemplateModel->template_id = $params['default_account_template_id'];
@@ -149,10 +149,11 @@ class AccountController extends \GO\Base\Controller\AbstractModelController {
 	}
 
 	protected function remoteComboFields() {
-		
+		if (\GO::modules()->addressbook)
 			return array('user_id' => '$model->user->name',
 					'default_template_id' => '$model->defaultTemplate->emailTemplate->name');
-		
+		else
+			return array('user_id' => '$model->user->name');
 	}
 
 	protected function actionCheckUnseen($params) {
@@ -262,6 +263,7 @@ class AccountController extends \GO\Base\Controller\AbstractModelController {
 
 	public function actionTree($params) {
 		\GO::session()->closeWriting();
+
 
 		$response = array();
